Fix many stuffs according to review comments
