+++ apps/gendsa.c
@@ -134,7 +134,7 @@ int gendsa_main(int argc, char **argv)
     DSA_free(dsa);
     release_engine(e);
     OPENSSL_free(passout);
-    bio_tempfile_cleanup();
+    apps_bf_tempfile_cleanup();
     return ret;
 }
 #endif
