+++ apps/pkcs8.c
@@ -354,7 +354,7 @@ int pkcs8_main(int argc, char **argv)
     BIO_free(in);
     OPENSSL_free(passin);
     OPENSSL_free(passout);
-    bio_tempfile_cleanup();
+    apps_bf_tempfile_cleanup();
 
     return ret;
 }
