+++ apps/pkey.c
@@ -230,7 +230,7 @@ int pkey_main(int argc, char **argv)
     BIO_free(in);
     OPENSSL_free(passin);
     OPENSSL_free(passout);
-    bio_tempfile_cleanup();
+    apps_bf_tempfile_cleanup();
 
     return ret;
 }
