+++ apps/req.c
@@ -883,7 +883,7 @@ int req_main(int argc, char **argv)
         OPENSSL_free(passin);
     if (passout != nofree_passout)
         OPENSSL_free(passout);
-    bio_tempfile_cleanup();
+    apps_bf_tempfile_cleanup();
     return ret;
 }
